Classify occurrence records in levels of confidence in species determination

Classifies occurrence records in levels of confidence in species determination

classify_occ(
  occ,
  spec = NULL,
  na.rm.coords = TRUE,
  crit.levels = c("det_by_spec", "taxonomist", "image", "sci_colection", "field_obs",
    "no_criteria_met"),
  ignore.det.names = NULL,
  spec.ambiguity = "not.spec",
  institution.source = "institutionCode",
  collection.code = "collectionCode",
  catalog.number = "catalogNumber",
  year.event = "year",
  date.identified = "dateIdentified",
  scientific.name = "species",
  determined.by = "identifiedBy",
  longitude = "decimalLongitude",
  latitude = "decimalLatitude",
  basis.of.rec = "basisOfRecord",
  media.type = "mediaType",
  occ.id = "occurrenceID"
)

Arguments

occ
Data frame with occurrence records information.

spec
Data frame with specialists' names. See details.

na.rm.coords
Logical. If TRUE, remove occurrences with NA in latitude or longitude

crit.levels
Character. Vector with levels of confidence in decreasing order. The criteria allowed are det_by_spec, taxonomist, image, sci_colection, field_obs, no_criteria_met. See details.

ignore.det.names
character vector indicatiing strings in the determined.by columns that should be ignored as a taxonomist. See details.

spec.ambiguity
character. Indicates how to deal with ambiguity in specialists names. not.spec solve ambiguity by classifing the determinating as done by a non-specialist;is.spec assumes the determination was done by a specialist; manual.check enables the user to manually check all ambiguos names. Defaut is not.spec.

institution.source
Column name of occ with the name of the institution that provided the data.

collection.code
Column name of occ with the codes for institution names.

catalog.number
Column name of occ with catalog number.

year.event
Column name of occ the year of the collection event.

date.identified
Column name of occ with the date of species determination.

scientific.name
Column name of occ with the species names.

determined.by
Column name of occ with the name of who determined the species.

longitude
Column name of occ longitude in decimal degrees.

latitude
Column name of occ latitude in decimal degrees.

basis.of.rec
Column name of occ with the recording types, as in GBIF.

media.type
Column name of occ with the media type of recording.

occ.id
Column name of occ with link or code for the occurence record.

Value

The occ data frame plus the classification of each record in a new column, named naturaList_levels.

Details

spec data frame must have columns separating LastName, Name and Abbrev. See create_spec_df function for a easy way to produce this data frame.

When ignore.det.name = NULL (default), the function ignores strings with only "RRC ID Flag", "NA", "", "-" and "_". When a character vector is provided, the function adds the strings in character vector to the default strings and ignore all these strings as being a name of a taxonomist.

basis.of.rec is a character vector with one of the following types of record:PRESERVED_SPECIMEN or HUMAN_OBSERVATION, as in GBIF data 'basisOfRecord'.

media.type uses the same pattern as GBIF mediaType column, indicating the existence of an associated image with STILLIMAGE.

See also

Author

Arthur V. Rodrigues

Examples

occ.class <- classify_occ(A.setosa, speciaLists)

#> Error in loadNamespace(j <- i[[1L]], c(lib.loc, .libPaths()), versionCheck = vI[[j]]): there is no package called 'Rcpp'

occ.class

#> Error in eval(expr, envir, enclos): objeto 'occ.class' não encontrado