Evaluate the cleaning of occurrences records

This function compare the area occupied by a species before and after pass through the cleaning procedure according to the chosen level of filter. The comparison can be made by measuring area in the geographical and in the environmental space

clean_eval(
  occ.cl,
  geo.space,
  env.space,
  level.filter = c("1_det_by_spec"),
  r,
  scientific.name = "species",
  longitude = "decimalLongitude",
  latitude = "decimalLatitude"
)

Arguments

occ.cl
Data frame with occurrence records information already classified by classify_occ function.

geo.space
a SpatialPolygons* or sf object defining the geographical space

env.space
a SpatialPolygons* or sf object defining the environmental space. Use the define_env_space for create this object.

level.filter
a character vector including the levels in 'naturaList_levels' collumn which filter the occurrence data set.

r
a raster with 2 layers representing the environmental variables

scientific.name
Column name of occ.cl with the species names.

longitude
Column name of occ.cl longitude in decimal degrees.

latitude
Column name of occ.cl latitude in decimal degrees.

Value

a list in which:

area data frame remaining area after cleaning proportional to the area before cleaning. The values vary from 0 to 1. Column named r.geo.area is the remaining area for all species in the geographic space and the r.env.area in the environmental space.

comp data frame with composition of species in sites (cells from raster layers) before cleaning (comp$comp$BC) and after cleaning (comp$comp$AC). The number of rows is equal the number of cells in r, and number of columns is equal to the number of species in the occ.cl.

rich data frame with a single column with the richness of each site

site.coords data frame with site's coordinates. It facilitates to built raster layers from results using rasterFromXYZ

See also

Examples

if (FALSE) {
pkgs <- c("devtools", "raster", "maptools", "rnaturalearth", "sf")
lapply(pkgs, require, character.only = TRUE)

data("A.setosa")
data("speciaLists")

# classify
occ.cl <- classify_occ(A.setosa, speciaLists, spec.ambiguity = "not.spec")
## check points
occ.cl <-map_module(occ.cl) #delete points in the ocean

# download climate data
bioclim <- getData('worldclim', var='bio', res=10)

# Transform occurrence data in SpatialPointsDataFrame
spdf.occ.cl <- SpatialPoints(occ.cl[, c("decimalLongitude", "decimalLatitude")])

# redefine the extent of bioclim layers based on buffer around the occurrences
c.bioclim <- crop(bioclim, buffer(spdf.occ.cl, 100000)) # 100km buffer

# select two layers
raster.temp.prec <- c.bioclim[[c("bio1", "bio12")]]
df.temp.prec <- as.data.frame(raster.temp.prec)

### Define the environmental space for analysis
env.space <- define_env_space(df.temp.prec, buffer.size = 0.05)

# delimit the geographic space
# land area
land <- ne_countries(continent = 'south america')
un.land <- unionSpatialPolygons(land, land$scalerank)

# geo space based on crop
c.geo.space <- crop(un.land, c.bioclim)

#geo space based on intersect
i.geo.space <- intersect(un.land, buffer(spdf.occ.cl2, 200000))

### run the evaluation
cl.eval <- clean_eval(occ.cl,
                      env.space = env.space,
                      geo.space = c.geo.space,
                      r = raster.temp.prec)

#area results
cl.eval$area

### richness maps
## it makes sense if there are more than one species
rich.before.clean <- rasterFromXYZ(cbind(cl.eval$site.coords,
                                         cl.eval$rich$rich.BC))
rich.after.clean <- rasterFromXYZ(cbind(cl.eval$site.coords,
                                        cl.eval$rich$rich.AC))

plot(rich.before.clean)
plot(rich.after.clean)

### species area map
comp.bc <- as.data.frame(cl.eval$comp$comp.BC)
comp.ac <- as.data.frame(cl.eval$comp$comp.AC)

c.setosa.bc <- rasterFromXYZ(cbind(cbind(cl.eval$site.coords,
                                         comp.bc$`Cyathea setosa`)))
c.setosa.ac <- rasterFromXYZ(cbind(cbind(cl.eval$site.coords,
                                         comp.ac$`Cyathea setosa`)))

plot(c.setosa.bc)
plot(c.setosa.ac)

}