Get the names in the 'identified.by' column

This function facilitates the search for non-taxonomist strings in the 'identified.by' column of occurrence records data set

get_det_names(
  occ,
  identified.by = "identifiedBy",
  freq = FALSE,
  decreasing = TRUE,
  determined.by
)

Arguments

occ
data frame with occurrence records information.

identified.by
column name of occ with the name of who determined the species.

freq
logical. If TRUE output contain the number of times each string is repeated in the identified.by column. Default = FALSE

decreasing
logical. sort strings in decreasing order of frequency. Default = TRUE.

determined.by
deprecated, use identified.by instead.

Value

character vector containing characters in identified.by column of occ.

Examples

data("A.setosa")
get_det_names(A.setosa, freq = TRUE)

#>                                           string frequency
#> 1                                                       79
#> 2                                   I. Fernandes        40
#> 3                                  Gasper, AL de        38
#> 4                                      Salino, A        26
#> 5                                  Schmitt, J.L.        17
#> 6                                    I.Fernandes        14
#> 7                                   Gonzatti, F.        13
#> 8                                  Fernandes, I.        11
#> 9                                   Ponce, M. M.        11
#> 10                                    R. Finotti         8
#> 11                                             -         7
#> 12                                   Gonzatti, F         7
#> 13                                    Salino, A.         7
#> 14                                     A. Salino         6
#> 15                                    Gasper, AL         6
#> 16                             Kieling-Rubio, MA         6
#> 17                             Rolf A. Engelmann         6
#> 18                                   Senna, R.M.         6
#> 19                                    C. R. Lehn         5
#> 20                                   Matos, F.B.         5
#> 21                                 A. Smith (UC)         4
#> 22                                    C. Mynssen         4
#> 23                                Canestraro, BK         4
#> 24                                  Labiak, P.H.         4
#> 25                            Schwartsburd, P.B.         4
#> 26                               Schwirkowski, P         4
#> 27                                      Verdi, M         4
#> 28                               Fernandes, I.M.         3
#> 29                                   J. Cordeiro         3
#> 30                             M. Lehnert (BONN)         3
#> 31                                     Matos, FB         3
#> 32                              Rolf A.Engelmann         3
#> 33                                Rolf Engelmann         3
#> 34                                     Senna, RM         3
#> 35                                 A.L.de Gasper         2
#> 36                                   C.E.Jascone         2
#> 37                                    Dreveck, S         2
#> 38                                   F. Gonzatti         2
#> 39                                      F. Matos         2
#> 40                                     Fleck, R.         2
#> 41                               Gasper, A.L. de         2
#> 42                      J. Cordeiro & E. Barboza         2
#> 43                                  J.L. Schmitt         2
#> 44                               Lehn & Steffens         2
#> 45                                         Lopes         2
#> 46                                  M.G. Caxambu         2
#> 47                              Mazziero, F.F.F.         2
#> 48                                   Michelon, C         2
#> 49                                 Nóbrega, G.A.         2
#> 50                                  P. H. Labiak         2
#> 51                                 P.G. Windisch         2
#> 52                               Pereira, J.B.S.         2
#> 53                                   RRC ID Flag         2
#> 54                                Sem Informação         2
#> 55                                    Záchia, R.         2
#> 56                                   A. Santiago         1
#> 57                                     A. Sehnem         1
#> 58                                      A. Smith         1
#> 59                             A. Smith, 1987-88         1
#> 60                                    A.C. Brade         1
#> 61                                A.L. de Gasper         1
#> 62                                    A.L.Gasper         1
#> 63                                   Almeida, TE         1
#> 64  Almeida, TE; Mota, NFO; Lau, BL; Ramalho, AJ         1
#> 65                                     Anonymous         1
#> 66                                 Arantes, A.A.         1
#> 67                                    Areias, G.         1
#> 68                      B. Zimmer (B) 2018-04-04         1
#> 69                                Barros, I.C.L.         1
#> 70                               Belgrano, M. J.         1
#> 71                                     Bueno, RM         1
#> 72                              Burmeister, E.L.         1
#> 73                      C. Mynssen et E. Moreira         1
#> 74                   C. R. Lehn & J. L. Schmitt.         1
#> 75                                   C. Steffens         1
#> 76                                  C.E. Jascone         1
#> 77                                    C.E.Jacone         1
#> 78                                   C.G.V.Ramos         1
#> 79                                     C.MYNSSEN         1
#> 80                              Canestraro, B.K.         1
#> 81                                  Carmes, A.A.         1
#> 82                                   Condack, J.         1
#> 83                              Dittrich, V.A.O.         1
#> 84                                 Dittrich, VAO         1
#> 85                             E. I. Meza Torres         1
#> 86                               E. R. Damasceno         1
#> 87                                   Estevan, DA         1
#> 88                                    F. Azevedo         1
#> 89                            F.S.Souza/Fev.2011         1
#> 90                                  Fernande, I.         1
#> 91                                   Funez, L.A.         1
#> 92                               G.A.D.C. Franco         1
#> 93                                    G.S. McKee         1
#> 94                                 Gasper, A. L.         1
#> 95                                  Gasper, A.L.         1
#> 96                      Goetz, MNB & Becker, DFP         1
#> 97                                Hiendlmayer, R         1
#> 98                                   I. Fernades         1
#> 99                                I. K. Flausino         1
#> 100                   I. K. Flausino/07-XII-2005         1
#> 101                                J. P. Condack         1
#> 102                                     J. Prado         1
#> 103                                   J. Schmitt         1
#> 104                                 Jarenkow, JA         1
#> 105                        Kieling- Rubio, M. A.         1
#> 106                         L.C. Bernacci et al.         1
#> 107                                  L.Sylvestre         1
#> 108                                Labiak, P. H.         1
#> 109                      Lana da Silva Sylvestre         1
#> 110                                 Lehnert (UC)         1
#> 111                                   Lermer, J.         1
#> 112                                   Lerner, J.         1
#> 113                               M.Lehnert (UC)         1
#> 114                                   M.M. Ponce         1
#> 115                               Marcus Lehnert         1
#> 116                                  Marquez, G.         1
#> 117                                    Matos, F.         1
#> 118                                Mckee, G. .S.         1
#> 119                         Melo, LCN; Salino, A         1
#> 120                                 Michelon, C.         1
#> 121                                Morero, R. E.         1
#> 122                                 Nervo, M. H.         1
#> 123                                    Nervo, MH         1
#> 124                              Oliveira, AA de         1
#> 125                                    P. Labiak         1
#> 126                           P.J.S. Silva Filho         1
#> 127                                        Ponce         1
#> 128                             Ponce, M. M., //         1
#> 129                                     Prado, J         1
#> 130                                 R. Engelmann         1
#> 131                                  R. M. Bueno         1
#> 132                                  R. M. Senna         1
#> 133                                    R.M.Senna         1
#> 134                            Rev. I. Fernandes         1
#> 135                                   Ristow, R.         1
#> 136                                Rogério Bueno         1
#> 137                                  Rubio, A.K.         1
#> 138                               S. Vasconcelos         1
#> 139                               Sakagami, C.R.         1
#> 140                                 Sakagami, CR         1
#> 141                        Salino, A; Nóbrega, T         1
#> 142                               Schmitt, J. L.         1
#> 143                                Schmitt. J.L.         1
#> 144                                    Silva, FD         1
#> 145                                 Silva, V. L.         1
#> 146                                 Steffens, C.         1
#> 147                                    Uller, HF         1
#> 148                V. Graeff, T. Padoin, V. Leão         1
#> 149                                  V. L. Silva         1
#> 150                              V.A.O. Dittrich         1
#> 151                    Vanilde Citadini- Zanette         1
#> 152                                    Verdi, M.         1
#> 153                                  Völtz, R.R.         1
#> 154                                 Windisch, PG         1