Check the occurrence records in a interactive map module

Allows to delete occurrence records and to select occurrence points by classification levels or by drawing spatial polygons.

map_module(
  occ.cl,
  action = "clean",
  institution.code = "institutionCode",
  collection.code = "collectionCode",
  catalog.number = "catalogNumber",
  year = "year",
  date.identified = "dateIdentified",
  species = "species",
  identified.by = "identifiedBy",
  decimal.latitude = "decimalLatitude",
  decimal.longitude = "decimalLongitude",
  basis.of.record = "basisOfRecord",
  media.type = "mediaType",
  occurrence.id = "occurrenceID",
  institution.source,
  year.event,
  scientific.name,
  determined.by,
  latitude,
  longitude,
  basis.of.rec,
  occ.id
)

Arguments

occ.cl
Data frame with occurrence records information already classified by classify_occ function.

action
a string with `"clean"` or `"flag"` which defines the action of `map_module` function with the occurrence dataset. Default is `"clean"`. If the string is `"clean"` the dataset returned only the occurrences records selected by the user. If the string is `"flag"`, a column named `map_module_flag` is added in the output dataset, with tags `selected` and `deleted`, following the choices of the user in the application.

institution.code
column name of occ with the name (or acronym) in use by the institution having custody of the object(s) or information referred to in the record.

collection.code
column name of occ with The name, acronym, code, or initials identifying the collection or data set from which the record was derived.

catalog.number
column name of occ with an identifier (preferably unique) for the record within the data set or collection.

year
Column name of occ the four-digit year in which the Event occurred, according to the Common Era Calendar.

date.identified
Column name of occ with the date on which the subject was determined as representing the Taxon.

species
column name of occ with the species names.

identified.by
column name of occ with the name of who determined the species.

decimal.latitude
column name of occ latitude in decimal degrees.

decimal.longitude
column name of occ longitude in decimal degrees.

basis.of.record
column name with the specific nature of the data record. See details.

media.type
column name of occ with the media type of recording. See details.

occurrence.id
column name of occ with link or code for the occurrence record. See in Darwin Core Format

institution.source
deprecated, use institution.code instead.

year.event
deprecated, use year instead.

scientific.name
deprecated, use species instead.

determined.by
deprecated, use identified.by instead

latitude
deprecated, use decimal.latitude instead

longitude
deprecated, use decimal.longitude instead

basis.of.rec
deprecated, use basis.of.record instead.

occ.id
deprecated, use occurrence.id instead

Value

Data frame with the same columns of occ.cl.

See also

Author

Arthur V. Rodrigues

Examples

if (FALSE) {
occ.class <- classify_occ(A.setosa, speciaLists)
occ.selected <- map_module(occ.class)
occ.selected
}